requirements.txt
pandas==1.5.3
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
python-dotenv==0.21.1